Delete AuthorizationCodeScopes
