+++ src/main/kotlin/com/marblet/idp/domain/model/AuthorizationCode.kt
@@ -7,7 +7,7 @@ class AuthorizationCode(
     val code: String,
     val userId: UserId,
     val clientId: ClientId,
-    val scopes: AuthorizationCodeScopes,
+    val scopes: ConsentedScopes,
     val redirectUri: RedirectUri,
     val expiration: LocalDateTime,
 ) {
@@ -27,34 +27,11 @@ class AuthorizationCode(
                     .map { Random.nextInt(0, charPool.size).let { charPool[it] } }
                     .joinToString("")
             val expiration = LocalDateTime.now().plusMinutes(EXPIRATION_MINUTES)
-            return AuthorizationCode(code, userId, clientId, scopes.toAuthorizationCodeScopes(), redirectUri, expiration)
+            return AuthorizationCode(code, userId, clientId, scopes, redirectUri, expiration)
         }
     }
 
     fun isExpired(): Boolean {
         return expiration.isBefore(LocalDateTime.now())
     }
-}
-
-/**
- * AuthorizationCodeScopesは、認可コードに持たせるOauth2.0文脈のスコープを表す。
- * "openid"やUserInfoのClaimsとして使われるスコープは、このスコープから除外する。
- */
-data class AuthorizationCodeScopes(private val value: Set<String>) {
-    companion object {
-        fun fromSpaceSeparatedString(scope: String) = AuthorizationCodeScopes(scope.split(" ").toSet())
-    }
-
-    fun toSpaceSeparatedString() = value.joinToString(" ")
-
-    fun toTokenScopes(): TokenScopes? {
-        // remove UserInfo scopes and openid scope
-        val scopes = value - (UserInfoScope.entries.map { it.value }.toSet() + OpenidScope.entries.map { it.value }.toSet())
-        if (scopes.isEmpty()) {
-            return null
-        }
-        return TokenScopes(scopes)
-    }
-
-    fun hasOpenidScope() = value.contains(OpenidScope.OPENID.value)
 }
