+++ src/test/kotlin/com/marblet/idp/domain/model/AuthorizationCodeTest.kt
@@ -44,57 +44,3 @@ class AuthorizationCodeTest {
         assertThat(actual.expiration).isBefore(expirationThreshold)
     }
 }
-
-class AuthorizationCodeScopesTest {
-    @Test
-    fun canGenerateFromSpaceSeparatedString() {
-        val actual = AuthorizationCodeScopes.fromSpaceSeparatedString("a b c")
-
-        assertThat(actual).isEqualTo(AuthorizationCodeScopes(setOf("a", "b", "c")))
-    }
-
-    @Test
-    fun canGenerateSpaceSeparatedString() {
-        val target = AuthorizationCodeScopes(setOf("a", "b", "c"))
-
-        val actual = target.toSpaceSeparatedString()
-
-        assertThat(actual).isEqualTo("a b c")
-    }
-
-    @Test
-    fun returnTrueIfScopesHasOpenidScope() {
-        val target = AuthorizationCodeScopes(setOf("openid", "email"))
-
-        val actual = target.hasOpenidScope()
-
-        assertThat(actual).isTrue()
-    }
-
-    @Test
-    fun returnFalseIfNoOpenidScope() {
-        val target = AuthorizationCodeScopes(setOf("a", "b"))
-
-        val actual = target.hasOpenidScope()
-
-        assertThat(actual).isFalse()
-    }
-
-    @Test
-    fun returnTokenScopeIfAccessTokenScopesExist() {
-        val target = AuthorizationCodeScopes(setOf("openid", "email", "a", "b"))
-
-        val actual = target.toTokenScopes()
-
-        assertThat(actual).isEqualTo(TokenScopes(setOf("a", "b")))
-    }
-
-    @Test
-    fun returnNullIfAccessTokenScopesNotExist() {
-        val target = AuthorizationCodeScopes(setOf("openid", "phone"))
-
-        val actual = target.toTokenScopes()
-
-        assertThat(actual).isNull()
-    }
-}
